added carousel component
