+++ client/src/pages/Home/Home.js
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import ButtonBtn from "../../components/ButtonBtn";
 import "./Home.css";
 import Nav from "../../components/Nav";
-import IndeedImg from '../../images/Indeed-Logo-1-1.jpg';
+import Carousel from "../../components/Carousel/Carousel"
 
 class Home extends Component {
   logout = () => {
@@ -13,50 +13,8 @@ class Home extends Component {
     window.location.reload();
   };
 
-  profilesList = () => {
-    let profiles = [
-      {
-        sourceImage: IndeedImg,
-        description: "Indeed Job Listing",
-        postUrl: "https://www.indeed.com"
-      },
-      {
-        sourceImage: IndeedImg,
-        description: "Built In Chicago Job Listing",
-        postUrl: "https://www.builtinchicago.org/"
-      },
-      {
-        sourceImage:
-        IndeedImg,
-        description: "Glassdoor Job Listing",
-        postUrl: "https://www.glassdoor.com/index.htm"
-      },
-      {
-        sourceImage: IndeedImg,
-        description: "Stackoverflow Job Listing",
-        postUrl: "https://stackoverflow.com/jobs?med=site-ui&ref=jobs-tab"
-      }
-    ];
-    return profiles.map(profile => (
-      <div
-        style={{ textAlign: "center", width: 370}}
-        className={
-          profiles[0] === profile ? "carousel-item active" : "carousel-item"
-        }
-      >
-        <img src={profile.sourceImage} alt="sourceImage"/>
-
-        <a href={profile.postUrl} target="_blank">
-          <div>{profile.description}</div>
-        </a>
-      </div>
-    ));
-  };
 
   render() {
-
-console.log('this is our pic we imported!!', IndeedImg);
-
 
     return (
       <div className="top-container">
@@ -83,45 +41,7 @@ console.log('this is our pic we imported!!', IndeedImg);
             {/*end five columns jumbo-text */}
 
             {/* <div className="one column">&nbsp;</div> */}
-            <div className="seven columns">
-              <div className="screenshot-container">
-                <div
-                  id="carouselExampleControls"
-                  className="carousel slide content"
-                  data-ride="carousel"
-                >
-                  <a
-                    className="carousel-control-prev"
-                    href="#carouselExampleControls"
-                    role="button"
-                    data-slide="prev"
-                  >
-                    <span
-                      className="carousel-control-prev-icon"
-                      aria-hidden="true"
-                    />
-                    <span className="sr-only">Previous</span>
-                  </a>
-                  <div className="carousel-inner">{this.profilesList()}</div>
-                  <a
-                    className="carousel-control-next"
-                    href="#carouselExampleControls"
-                    role="button"
-                    data-slide="next"
-                  >
-                    <span
-                      className="carousel-control-next-icon"
-                      aria-hidden="true"
-                    />
-                    <span className="sr-only">Next</span>
-                  </a>
-                </div> 
-                    {/*end carousel slide content*/}
-                <script>$('.carousel').carousel();</script>
-              </div>
-                {/*end screenshot-container */}
-            </div>
-                {/* end seven columns menu */}
+            <Carousel/>
           </div>
             {/* end row*/}
         </div>
